Summarizing Generalized Linear Model Fits with FAB Inference

summary method for class glmFAB

# S3 method for glmFAB
summary(
  object,
  dispersion = NULL,
  correlation = FALSE,
  symbolic.cor = FALSE,
  ...
)

Arguments

object
an object of class glmFAB

dispersion
see summary.glm

correlation
see summary.glm

symbolic.cor
see summary.glm

...
see summary.glm

Value

A list of summary statistics of the fitted generalized linear model

Details

A mod of summary.glm that shows FAB p-values in table

Examples

# n observations, p FAB variables, q=2 control variables 

n<-100 ; p<-25 

# X is design matrix for params of interest
# beta is vector of true parameter values 
# v a variable in the linking model - used to share info across betas

v<-rnorm(p) ; beta<-(2 - 2*v + rnorm(p))/3 ; X<-matrix(rnorm(n*p),n,p)/8

# control coefficients and variables  
alpha1<-.5 ; alpha2<- -.5
w1<-rnorm(n)/8
w2<-rnorm(n)/8

# simulate data 
lp<-1 + alpha1*w1 + alpha2*w2 + X%*%beta 
y<-rpois(n,exp(lp))

# fit model
fit<-glmFAB(y~w1+w2,X,~v,family=poisson)

#> 
#> Fitting sampling model: -  Fitting sampling model: #  
#>  Fitting linking models: #------------------------ Fitting linking models: ##----------------------- Fitting linking models: ###---------------------- Fitting linking models: ####--------------------- Fitting linking models: #####-------------------- Fitting linking models: ######------------------- Fitting linking models: #######------------------ Fitting linking models: ########----------------- Fitting linking models: #########---------------- Fitting linking models: ##########--------------- Fitting linking models: ###########-------------- Fitting linking models: ############------------- Fitting linking models: #############------------ Fitting linking models: ##############----------- Fitting linking models: ###############---------- Fitting linking models: ################--------- Fitting linking models: #################-------- Fitting linking models: ##################------- Fitting linking models: ###################------ Fitting linking models: ####################----- Fitting linking models: #####################---- Fitting linking models: ######################--- Fitting linking models: #######################-- Fitting linking models: ########################- Fitting linking models: #########################

fit$FABpv

#>          fv1          fv2          fv3          fv4          fv5          fv6 
#> 3.301144e-02 9.911499e-01 2.861218e-01 3.501581e-01 4.298153e-01 1.294312e-01 
#>          fv7          fv8          fv9         fv10         fv11         fv12 
#> 3.484147e-02 2.908584e-02 1.418045e-01 7.705731e-01 4.051792e-03 7.839975e-01 
#>         fv13         fv14         fv15         fv16         fv17         fv18 
#> 2.985106e-01 2.235317e-01 1.021301e-03 6.531914e-01 1.272534e-02 1.886381e-03 
#>         fv19         fv20         fv21         fv22         fv23         fv24 
#> 2.076070e-03 7.617339e-02 3.158509e-03 8.761060e-05 6.959489e-01 1.838333e-01 
#>         fv25 
#> 5.882124e-08 

fit$FABci 

#>            [,1]        [,2]
#> fv1   0.1142576  2.05898805
#> fv2  -1.9878811  0.20317560
#> fv3  -0.5918953  1.21079061
#> fv4  -0.7445656  1.19953879
#> fv5  -0.7867429  0.97627772
#> fv6  -0.3189098  1.71484591
#> fv7  -2.4743405 -0.06418161
#> fv8   0.1421588  2.01724496
#> fv9  -0.3267647  1.55028486
#> fv10 -1.2520181  0.48645769
#> fv11  0.5419691  2.32013537
#> fv12 -1.3668323  0.48314354
#> fv13 -0.7377862  1.43663812
#> fv14 -0.5114647  1.39081652
#> fv15  0.8010158  2.63237019
#> fv16 -1.2538535  1.35575830
#> fv17  0.2787870  1.83506839
#> fv18  0.7585767  2.75259070
#> fv19  0.6532015  2.41267292
#> fv20 -0.1384665  1.99340604
#> fv21  0.2956332  2.72607958
#> fv22  1.3294796  3.96144926
#> fv23 -0.6269868  1.19492796
#> fv24 -0.5162242  1.76613042
#> fv25  1.6665975  3.16807513

summary(fit) # look at p-value column 

#> 
#> Call:
#> glm(formula = y ~ . + 0, family = ..1, data = as.data.frame(cbind(W, 
#>     X)))
#> 
#> Deviance Residuals: 
#>      Min        1Q    Median        3Q       Max  
#> -2.55395  -0.87686  -0.05881   0.49388   2.00750  
#> 
#> Coefficients:
#>               Estimate Std. Error z value Pr(>|z+bfab|)    
#> `(Intercept)`  0.94317    0.07717  12.222       < 2e-16 ***
#> w1             0.83149    0.61437   1.353       0.17593    
#> w2            -0.74838    0.57911  -1.292       0.19626    
#> fv1            1.08662    0.59111   1.838       0.03301 *  
#> fv2           -1.17376    0.49488  -2.372       0.99115    
#> fv3            0.30945    0.54794   0.565       0.28612    
#> fv4            0.22745    0.59094   0.385       0.35016    
#> fv5            0.09477    0.53588   0.177       0.42982    
#> fv6            0.69797    0.61817   1.129       0.12943    
#> fv7           -1.29761    0.71535  -1.814       0.03484 *  
#> fv8            1.07970    0.56994   1.894       0.02909 *  
#> fv9            0.61176    0.57054   1.072       0.14180    
#> fv10          -0.38278    0.52842  -0.724       0.77057    
#> fv11           1.43105    0.54048   2.648       0.00405 ** 
#> fv12          -0.44184    0.56231  -0.786       0.78400    
#> fv13           0.34943    0.66093   0.529       0.29851    
#> fv14           0.43964    0.57823   0.760       0.22353    
#> fv15           1.71669    0.55665   3.084       0.00102 ** 
#> fv16          -0.24226    0.61496  -0.394       0.65319    
#> fv17           1.05693    0.47301   2.234       0.01273 *  
#> fv18           1.75558    0.60609   2.897       0.00189 ** 
#> fv19           1.53294    0.53480   2.866       0.00208 ** 
#> fv20           0.92747    0.64800   1.431       0.07617 .  
#> fv21           1.70128    0.62299   2.731       0.00316 ** 
#> fv22           2.75408    0.73397   3.752      8.76e-05 ***
#> fv23           0.28397    0.55378   0.513       0.69595    
#> fv24           0.62495    0.69374   0.901       0.18383    
#> fv25           2.41734    0.45635   5.297      5.88e-08 ***
#> ---
#> Signif. codes:  0 ‘***’ 0.001 ‘**’ 0.01 ‘*’ 0.05 ‘.’ 0.1 ‘ ’ 1
#> 
#> (Dispersion parameter for poisson family taken to be 1)
#> 
#>     Null deviance: 601.960  on 100  degrees of freedom
#> Residual deviance:  99.218  on  72  degrees of freedom
#> AIC: 405.32
#> 
#> Number of Fisher Scoring iterations: 5
#>